+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed arrow direction in `util\network\addToNetwork.m` and changed parameter name from `useWeights` to `weighted` in `util\network\plotNetwork.m`
 - Added utility functions: `util\plots\lighten.m`, `util\analysis\helpers\printOffEdges.m`.
 - Fixed `util\analysis\iplDepth.m` for InferiorMonkey
+- More informative error messages in `views\RenderApp.m`
 
 ### 18Oct2018
 - Fixed issue with `RenderApp.m` UI layout for NeitzTemporalMonkey and MarcRC1